refactored Vue components into folders
