+++ extension/src/components/modals/ConfirmModal.vue
