+++ extension/src/components/modals/DAOConfigModal.vue
