+++ extension/src/components/modals/VotingModal.vue
